Subject(s) of each query are converted into a **frozenset**. This is because frozensets are _hashable_, a property necessary for subsequent assignment result caching.

Memory footprint increases because of the copy of this large data structure, but it is trivial compared with the overall memory consumption of the program. It is an option to return a generator instead of the actual data, which will be faster and less memory-consuming. However, the obstacle is that one or multiple ranks may be specified by the user, and this data structure needs to be re-used in the latter case. Therefore, a generator won't work.

This change further allows the plain mapper to yield queues of **lists** instead of **sets**. Moderate performance improvement was observed along with this change:

- 214 ms ± 1.5 ms => 207 ms ± 1.52 ms