@qiyunzhu 2020-06-27
Optimized function `strip_index`, replacing nested list comprehensions with nested `map` functions.

```python
def before(subque, sep='_'):
    return (frozenset(x.rsplit(sep, 1)[0] for x in subs) for subs in subque)

def after(subque, sep='_'):
    return map(frozenset, map(partial(map, lambda x: x.rsplit(sep, 1)[0]), subque))
```

Benchmarks on a Bowtie2 alignment file with 968,054 lines:

- Before: 656 ms ± 10.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- After: 581 ms ± 12.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

Also tested multiple other options, including nested lambda functions, methodcaller, etc. The current solution is the best.